move integration tests for psycopg_async_notify into a single test file
